% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jettel.r
\name{checkStatus}
\alias{checkStatus}
\title{Check of accessibility of server}
\usage{
checkStatus(token = 0)
}
\arguments{
\item{token}{User token, string, default null.}
}
\description{
This function checks if Treffle API server is accessible
}
\examples{
checkStatus()
}
\keyword{server}
