% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jettel.r
\name{isToken}
\alias{isToken}
\title{Check Token}
\usage{
isToken(token = 0)
}
\arguments{
\item{token}{User token, string, default null.}
}
\description{
This function checks if user have provided token, if no site for obtaining token show up
}
\examples{
isToken()
}
\keyword{token}
