% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jettel.r
\name{retrieve.plant}
\alias{retrieve.plant}
\title{Get dataframe of only one plant selected form searched plants}
\usage{
retrieve.plant(token = 0, plant = "")
}
\arguments{
\item{token}{User token, string, default null.}

\item{plant}{User query, string, default null.}
}
\value{
A dataframe of selected plant, from list all plants which satisfy given search criteria
}
\description{
This function search plant which satisfies given query and lets user choose one plant.
}
\examples{
retrieve.plant("token", "coconut")
}
\keyword{plant}
\keyword{search}
