% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jettel.r
\name{search.kingdom}
\alias{search.kingdom}
\title{Get all kingdoms}
\usage{
search.kingdom(token = 0)
}
\arguments{
\item{token}{User token, string, default null.}
}
\value{
A dataframe of kingdoms
}
\description{
This function retrieves kingdoms
}
\examples{
search.kingdom("token")
}
\keyword{kingdom}
\keyword{search}
