% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jettel.r
\name{search.plant}
\alias{search.plant}
\title{Get all plants which satisfy given search criteria}
\usage{
search.plant(token = 0, plant = 0)
}
\arguments{
\item{token}{User token, string, default null.}

\item{plant}{User query, string, default null.}
}
\value{
A dataframe with plants which satisfy given search criteria
}
\description{
This function search plant which satisfies given query
}
\examples{
search.plant("token", "coconut")
}
\keyword{plant}
\keyword{search}
