% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jettel.r
\name{search.species}
\alias{search.species}
\title{Get all species which satisfy given search criteria}
\usage{
search.species(token = 0, species = "", page = 1)
}
\arguments{
\item{token}{User token, string, default null.}

\item{species}{User query, string, default null.}

\item{page}{Number of page, integer, default 1.}
}
\value{
A dataframe with species which satisfy given search criteria
}
\description{
This function search species which satisfies given query
}
\examples{
search.plant("token", "coconut")
}
\keyword{plant}
\keyword{search}
